refactor: split CLI audit and identity helpers
